+++ packages/nano_domain/lib/src/nano_domain.dart
@@ -6,6 +6,7 @@ export 'environment/service_endpoint.dart';
 export 'home/student_home_summary.dart';
 export 'profile/student_profile_view.dart';
 export 'learning/learning_catalog.dart';
+export 'learning/topic_action.dart';
 export 'learning/learning_subject.dart';
 export 'learning/home_plan_item.dart';
 export 'navigation/app_role.dart';